Fix type issue.
2003-12-04  Alexander Larsson  <alexl@redhat.com>

	* gnome-keyring.c (operation_io):
	Fix type issue.

+++ gnome-keyring.c
@@ -301,7 +301,7 @@ operation_io (GIOChannel  *io_channel,
 	GIOChannel *channel;
 	GnomeKeyringOperation *op;
 	int res;
-	gsize packet_size;
+	guint32 packet_size;
 
 	op = callback_data;
